+++ domain/src/main/java/com/upf464/koonsdiary/domain/di/UserUseCaseModule.kt
@@ -2,7 +2,8 @@ package com.upf464.koonsdiary.domain.di
 
 import com.upf464.koonsdiary.domain.request.*
 import com.upf464.koonsdiary.domain.response.EmptyResponse
-import com.upf464.koonsdiary.domain.usecase.*
+import com.upf464.koonsdiary.domain.usecase.ResultUseCase
+import com.upf464.koonsdiary.domain.usecase.user.*
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
