+++ domain/src/main/java/com/upf464/koonsdiary/domain/usecase/diary/AddDiaryUseCase.kt
@@ -1,34 +1,34 @@
-package com.upf464.koonsdiary.domain.usecase
+package com.upf464.koonsdiary.domain.usecase.diary
 
 import com.upf464.koonsdiary.domain.common.DiaryValidator
 import com.upf464.koonsdiary.domain.common.flatMap
 import com.upf464.koonsdiary.domain.error.DiaryError
 import com.upf464.koonsdiary.domain.model.Diary
 import com.upf464.koonsdiary.domain.repository.DiaryRepository
-import com.upf464.koonsdiary.domain.request.UpdateDiaryRequest
-import com.upf464.koonsdiary.domain.response.UpdateDiaryResponse
+import com.upf464.koonsdiary.domain.request.AddDiaryRequest
+import com.upf464.koonsdiary.domain.response.AddDiaryResponse
+import com.upf464.koonsdiary.domain.usecase.ResultUseCase
 import javax.inject.Inject
 
-internal class UpdateDiaryUseCase @Inject constructor(
-    private val diaryRepository: DiaryRepository,
-    private val validator: DiaryValidator
-) : ResultUseCase<UpdateDiaryRequest, UpdateDiaryResponse> {
+internal class AddDiaryUseCase @Inject constructor(
+    private val validator: DiaryValidator,
+    private val diaryRepository: DiaryRepository
+) : ResultUseCase<AddDiaryRequest, AddDiaryResponse> {
 
-    override suspend fun invoke(request: UpdateDiaryRequest): Result<UpdateDiaryResponse> {
+    override suspend fun invoke(request: AddDiaryRequest): Result<AddDiaryResponse> {
         if (!validator.validateContent(request.content)) {
             return Result.failure(DiaryError.EmptyContent)
         }
 
         val diary = Diary(
-            id = request.diaryId,
             date = request.date,
             content = request.content,
             sentiment = request.sentiment,
             imageList = request.imageList
         )
 
-        return diaryRepository.updateDiary(diary).flatMap { diaryId ->
-            Result.success(UpdateDiaryResponse(diaryId))
+        return diaryRepository.addDiary(diary).flatMap { diaryId ->
+            Result.success(AddDiaryResponse(diaryId))
         }
     }
 }
