+++ domain/src/main/java/com/upf464/koonsdiary/domain/usecase/diary/AnalyzeSentimentUseCase.kt
@@ -1,4 +1,4 @@
-package com.upf464.koonsdiary.domain.usecase
+package com.upf464.koonsdiary.domain.usecase.diary
 
 import com.upf464.koonsdiary.domain.common.DiaryValidator
 import com.upf464.koonsdiary.domain.common.flatMap
@@ -6,6 +6,7 @@ import com.upf464.koonsdiary.domain.error.DiaryError
 import com.upf464.koonsdiary.domain.repository.DiaryRepository
 import com.upf464.koonsdiary.domain.request.AnalyzeSentimentRequest
 import com.upf464.koonsdiary.domain.response.AnalyzeSentimentResponse
+import com.upf464.koonsdiary.domain.usecase.ResultUseCase
 import javax.inject.Inject
 
 internal class AnalyzeSentimentUseCase @Inject constructor(
