+++ domain/src/main/java/com/upf464/koonsdiary/domain/usecase/user/SignInWithUsernameUseCase.kt
@@ -1,10 +1,11 @@
-package com.upf464.koonsdiary.domain.usecase
+package com.upf464.koonsdiary.domain.usecase.user
 
 import com.upf464.koonsdiary.domain.common.HashGenerator
 import com.upf464.koonsdiary.domain.common.flatMap
 import com.upf464.koonsdiary.domain.repository.UserRepository
 import com.upf464.koonsdiary.domain.request.SignInWithUsernameRequest
 import com.upf464.koonsdiary.domain.response.EmptyResponse
+import com.upf464.koonsdiary.domain.usecase.ResultUseCase
 import javax.inject.Inject
 
 internal class SignInWithUsernameUseCase @Inject constructor(
