+++ domain/src/main/java/com/upf464/koonsdiary/domain/usecase/user/SignUpWithKakaoUseCase.kt
@@ -1,4 +1,4 @@
-package com.upf464.koonsdiary.domain.usecase
+package com.upf464.koonsdiary.domain.usecase.user
 
 import com.upf464.koonsdiary.domain.common.flatMap
 import com.upf464.koonsdiary.domain.common.handleWith
@@ -7,6 +7,7 @@ import com.upf464.koonsdiary.domain.repository.UserRepository
 import com.upf464.koonsdiary.domain.request.SignUpWithKakaoRequest
 import com.upf464.koonsdiary.domain.response.EmptyResponse
 import com.upf464.koonsdiary.domain.service.KakaoService
+import com.upf464.koonsdiary.domain.usecase.ResultUseCase
 import javax.inject.Inject
 
 internal class SignUpWithKakaoUseCase @Inject constructor(
