+++ domain/src/test/java/com/upf464/koonsdiary/domain/usecase/diary/AddDiaryUseCaseTest.kt
@@ -1,10 +1,10 @@
-package com.upf464.koonsdiary.domain.usecase
+package com.upf464.koonsdiary.domain.usecase.diary
 
 import com.upf464.koonsdiary.domain.common.DiaryValidator
 import com.upf464.koonsdiary.domain.error.DiaryError
 import com.upf464.koonsdiary.domain.model.Sentiment
 import com.upf464.koonsdiary.domain.repository.DiaryRepository
-import com.upf464.koonsdiary.domain.request.AnalyzeSentimentRequest
+import com.upf464.koonsdiary.domain.request.AddDiaryRequest
 import io.mockk.coEvery
 import io.mockk.every
 import io.mockk.mockkClass
@@ -12,18 +12,19 @@ import kotlinx.coroutines.runBlocking
 import org.junit.Assert.*
 import org.junit.Before
 import org.junit.Test
+import java.time.LocalDate
 
-class AnalyzeSentimentUseCaseTest {
+class AddDiaryUseCaseTest {
 
     private lateinit var validator: DiaryValidator
     private lateinit var diaryRepository: DiaryRepository
-    private lateinit var useCase: AnalyzeSentimentUseCase
+    private lateinit var useCase: AddDiaryUseCase
 
     @Before
     fun setup() {
         validator = mockkClass(DiaryValidator::class)
         diaryRepository = mockkClass(DiaryRepository::class)
-        useCase = AnalyzeSentimentUseCase(
+        useCase = AddDiaryUseCase(
             validator = validator,
             diaryRepository = diaryRepository
         )
@@ -32,19 +33,22 @@ class AnalyzeSentimentUseCaseTest {
     @Test
     fun invoke_validDiary_isSuccess(): Unit = runBlocking {
         coEvery {
-            diaryRepository.fetchSentimentOf("content")
-        } returns Result.success(Sentiment.GOOD)
+            diaryRepository.addDiary(any())
+        } returns Result.success(1)
 
         every {
             validator.validateContent("content")
         } returns true
 
-        val result = useCase(
-            AnalyzeSentimentRequest("content")
-        )
+        val result = useCase(AddDiaryRequest(
+            date = LocalDate.of(2022, 3, 25),
+            content = "content",
+            sentiment = Sentiment.GOOD,
+            imageList = emptyList()
+        ))
 
         assertTrue(result.isSuccess)
-        assertEquals(Sentiment.GOOD, result.getOrNull()?.sentiment)
+        assertEquals(1, result.getOrNull()?.diaryId)
     }
 
     @Test
@@ -53,9 +57,12 @@ class AnalyzeSentimentUseCaseTest {
             validator.validateContent("")
         } returns false
 
-        val result = useCase(
-            AnalyzeSentimentRequest("")
-        )
+        val result = useCase(AddDiaryRequest(
+            date = LocalDate.of(2022, 3, 25),
+            content = "",
+            sentiment = Sentiment.GOOD,
+            imageList = emptyList()
+        ))
 
         assertFalse(result.isSuccess)
         assertEquals(DiaryError.EmptyContent, result.exceptionOrNull())
